MGOGRelationship Dao, Manager and Controller finished Added "associated goal" into ApplicationResources GoalForm page now shows a select box for entering the associated goal The GoalController still doesn't support the "associated goal" field (it doens't fill the table mg_og_relationship)

+++ src/main/java/it/uniroma2/gqm/model/GoalType.java
@@ -26,6 +26,6 @@ public enum GoalType {
 	}
 	
 	public static boolean isMG(Goal g) {
-		return g.getType() == GoalType.MG.getId();
+		return g.getType().intValue() == GoalType.MG.getId();
 	}
 }
